requirements.txt
ray[serve]==2.9.0
torch==2.0.0
transformers==4.36.0
ffmpeg-python==0.2.0
pyyaml==6.0.1
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
pytest==7.4.3

